requirements.txt
bosdyn-client
bosdyn-mission
bosdyn-choreography-client
scipy